feat: impove build
