+++ README.md
@@ -1,7 +1,14 @@
 # meetups-ui
 Frontend for Meetups app
 
-# React + TypeScript + Vite
+## Before starting:
+
+Add a .env file with two lines following the example from the .env.example file.\
+You can find the data to fill out in the documentation in Clickup.
+
+------
+
+## React + TypeScript + Vite
 
 This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
 
@@ -10,7 +17,7 @@ Currently, two official plugins are available:
 - [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
 - [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
 
-## Expanding the ESLint configuration
+### Expanding the ESLint configuration
 
 If you are developing a production application, we recommend updating the configuration to enable type aware lint rules:
