+++ docker-compose.yaml
@@ -1,0 +1,31 @@
+version: '3'
+
+services:
+  # Development service
+  dev:
+    container_name: react-app-dev
+    build:
+      context: .
+      dockerfile: Dockerfile.dev
+    ports:
+      - "3000:3000"
+    volumes:
+      - ./src:/usr/src/app/src
+      - ./public:/usr/src/app/public
+    environment:
+      NODE_ENV: development
+    env_file:
+      - .env
+
+  # Production like service
+  prod:
+    container_name: react-app-prod
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - "80:80"
+    environment:
+      NODE_ENV: production
+    env_file:
+      - .env
