+++ package.json
@@ -10,18 +10,49 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@adobe/react-spectrum": "^3.32.2",
+    "@googlemaps/react-wrapper": "^1.1.35",
+    "@heroicons/react": "^1.0.6",
+    "@internationalized/date": "^3.5.0",
+    "@reduxjs/toolkit": "^1.9.7",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^20.11.0",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "@vis.gl/react-google-maps": "^0.3.3",
+    "axios": "^1.6.2",
+    "dotenv": "^16.3.1",
+    "env-cmd": "^10.1.0",
+    "lodash": "^4.17.21",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-final-form": "^6.5.9",
+    "react-hook-form": "^7.49.2",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.18.0",
+    "react-toastify": "^9.1.3",
+    "react-transition-group": "^4.4.5",
+    "sass": "^1.69.5",
+    "styled-components": "^6.1.1",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
+    "@types/lodash": "^4.14.202",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
+    "@types/react-redux": "^7.1.32",
+    "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
     "@typescript-eslint/parser": "^6.14.0",
     "@vitejs/plugin-react": "^4.2.1",
+    "dotenv-webpack": "^8.0.1",
     "eslint": "^8.55.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "tailwindcss": "^3.3.6",
     "typescript": "^5.2.2",
     "vite": "^5.0.8"
   }
