+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+import App from './app/app.tsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
